asdasdasdasd fix
Sometimes when I test things I will only try adasdasdasdasd during completion of this phrase if the output of the categorical interrogation is none then we oviouslt get none for the edit_attr action during runtime.  Well that breaks the prompt action that follows when querying from the pinecone namespace.  Created a default set_attr_safe which requires a truthy value and a value of the same type or the function will use the default from the endpoints.yaml